+++ tests/test_glm_hmm_validator.py
@@ -6,11 +6,8 @@ No model construction needed — validator methods are called directly on
 ``test_glm_hmm_class.py``.
 """
 
-from contextlib import nullcontext as does_not_raise
-
 import jax.numpy as jnp
 import numpy as np
-import pynapple as nap
 import pytest
 
 from nemos.glm_hmm.params import GLMHMMParams
@@ -192,166 +189,6 @@ class TestCheckInitAndTransitionProbSumTo1:
         params = (*valid_user_params[:4], bad_trans)
         with pytest.raises(ValueError):
             validator.check_init_and_transition_prob_sum_to_1(params)
-
-
-# ---------------------------------------------------------------------------
-# validate_inputs
-# Moved from test_glm_hmm_class.py::TestGLMHMM::test_nan_between_epochs
-# ---------------------------------------------------------------------------
-
-
-class TestValidateInputs:
-    """NaN validation: NaNs at epoch boundaries are allowed, NaNs in the middle
-    are rejected. Also covers X/y dimensionality and sample-count consistency."""
-
-    @pytest.mark.parametrize(
-        "X_ndim, expectation",
-        [
-            (1, pytest.raises(ValueError, match="X must be 2-dimensional")),
-            (2, does_not_raise()),
-            (3, pytest.raises(ValueError, match="X must be 2-dimensional")),
-        ],
-    )
-    def test_X_wrong_ndim_raises(self, validator, X_ndim, expectation):
-        n = 5
-        shape = {1: (n,), 2: (n, 1), 3: (n, 1, 1)}[X_ndim]
-        X = np.ones(shape)
-        y = np.zeros(n)
-        with expectation:
-            validator.validate_inputs(X, y)
-
-    @pytest.mark.parametrize(
-        "y_ndim, expectation",
-        [
-            (2, pytest.raises(ValueError, match="y must be 1-dimensional")),
-            (1, does_not_raise()),
-        ],
-    )
-    def test_y_wrong_ndim_raises(self, validator, y_ndim, expectation):
-        n = 5
-        X = np.ones((n, 1))
-        y = np.zeros((n, 1)) if y_ndim == 2 else np.zeros(n)
-        with expectation:
-            validator.validate_inputs(X, y)
-
-    def test_X_y_sample_mismatch_raises(self, validator):
-        with pytest.raises(ValueError, match="same number of samples"):
-            validator.validate_inputs(np.ones((5, 1)), np.zeros(6))
-
-    @pytest.mark.parametrize(
-        "X, y, expectation",
-        [
-            # NaN at start/end of array — allowed (epoch boundary)
-            (np.array([[np.nan], [0]]), np.array([0, 1]), does_not_raise()),
-            (np.array([[0], [np.nan]]), np.array([0, 1]), does_not_raise()),
-            (np.array([[0], [0]]), np.array([np.nan, 1]), does_not_raise()),
-            (np.array([[0], [0]]), np.array([0, np.nan]), does_not_raise()),
-            # NaN in the middle — rejected
-            (
-                np.array([[0], [np.nan], [0]]),
-                np.array([0, 1, 2]),
-                pytest.raises(ValueError, match="requires continuous time-series data"),
-            ),
-            (
-                np.array([[0], [0], [0]]),
-                np.array([0, np.nan, 2]),
-                pytest.raises(ValueError, match="requires continuous time-series data"),
-            ),
-            # Pynapple: NaN inside an epoch — rejected
-            (
-                nap.TsdFrame(
-                    t=np.arange(5),
-                    d=np.array([[0], [np.nan], [0], [0], [0]]),
-                    time_support=nap.IntervalSet([0, 3], [2, 5]),
-                ),
-                np.array([0, 1, 2, 4, 5]),
-                pytest.raises(ValueError, match="requires continuous time-series data"),
-            ),
-            # Pynapple: NaN at epoch boundary — allowed
-            (
-                nap.TsdFrame(
-                    t=np.arange(5),
-                    d=np.array([[0], [0], [np.nan], [0], [0]]),
-                    time_support=nap.IntervalSet([0, 3], [2, 5]),
-                ),
-                np.array([0, 1, 2, 4, 5]),
-                does_not_raise(),
-            ),
-            # Pynapple y: NaN inside epoch — rejected
-            (
-                np.zeros((5, 1)),
-                nap.Tsd(
-                    t=np.arange(5),
-                    d=np.array([0, np.nan, 2, 4, 5]),
-                    time_support=nap.IntervalSet([0, 3], [2, 5]),
-                ),
-                pytest.raises(ValueError, match="requires continuous time-series data"),
-            ),
-            # Pynapple y: NaN at epoch boundary — allowed
-            (
-                np.zeros((5, 1)),
-                nap.Tsd(
-                    t=np.arange(5),
-                    d=np.array([0, 1, np.nan, 4, 5]),
-                    time_support=nap.IntervalSet([0, 3], [2, 5]),
-                ),
-                does_not_raise(),
-            ),
-            # Multiple consecutive NaNs in the middle — rejected
-            (
-                np.array([[0], [np.nan], [np.nan], [0]]),
-                np.array([0, 1, 2, 3]),
-                pytest.raises(ValueError, match="requires continuous time-series data"),
-            ),
-            # Multiple consecutive NaNs at start — allowed
-            (
-                np.array([[np.nan], [np.nan], [0]]),
-                np.array([0, 1, 2]),
-                does_not_raise(),
-            ),
-            # Multiple consecutive NaNs at end — allowed
-            (
-                np.array([[0], [np.nan], [np.nan]]),
-                np.array([0, 1, 2]),
-                does_not_raise(),
-            ),
-            # All NaN — rejected (caught by parent validation)
-            (
-                np.array([[np.nan], [np.nan]]),
-                np.array([np.nan, np.nan]),
-                pytest.raises(ValueError),
-            ),
-            # No NaN — allowed
-            (np.array([[0], [1]]), np.array([0, 1]), does_not_raise()),
-            # Pynapple: NaN at start of second epoch — allowed
-            (
-                nap.TsdFrame(
-                    t=np.arange(5),
-                    d=np.array([[0], [0], [np.nan], [0], [0]]),
-                    time_support=nap.IntervalSet([0, 2], [2, 5]),
-                ),
-                np.zeros(5),
-                does_not_raise(),
-            ),
-            # Both X and y NaN in middle at different positions — rejected
-            (
-                np.array([[0], [np.nan], [0], [0]]),
-                np.array([0, 1, np.nan, 3]),
-                pytest.raises(ValueError, match="requires continuous time-series data"),
-            ),
-            # Both X and y NaN in middle at same position — rejected
-            (
-                np.array([[0], [np.nan], [0]]),
-                np.array([0, np.nan, 2]),
-                pytest.raises(ValueError, match="requires continuous time-series data"),
-            ),
-        ],
-    )
-    def test_nan_at_boundary_allowed_in_middle_rejected(
-        self, validator, X, y, expectation
-    ):
-        with expectation:
-            validator.validate_inputs(X, y)
 
 
 # ---------------------------------------------------------------------------
